Logger is adjusted
 - custom logger handler is added to the plugin module
