+++ pytest_reportportal/pytest_rp_plugin.py
@@ -5,6 +5,23 @@ import logging
 from service import PyTestService
 
 
+class RPlogHandler(logging.Handler):
+    def __init__(self, level="INFO"):
+
+        super(RPlogHandler, self).__init__(level)
+
+    def emit(self, record):
+        try:
+            msg = self.format(record)
+        except (KeyboardInterrupt, SystemExit):
+            raise
+        except:
+            self.handleError(record)
+
+        print("LOGLEVEL in RPlogHandler %s" % self.level)
+        return PyTestService.post_log(msg, log_level=self.level)
+
+
 class RP_Report_Listener(object):
 
     # Identifier if TestItem is called:
